Corrected issues to more-about banner and mobile sizing fixes
more-about banner had a broken link due to updated image name.

When flexing the screen from full to mobile the selected content outside of the Our Story section, would have the regular size content and the mobile buttons. While many users may not notice this, I adjusted so that when flexed to mobile sizing only mobile buttons and banner will show regardless of section you are on in desktop sizing.